pyproject.toml
[project]
name = "styl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv-python>=0.0.1",
    "google-genai>=1.16.1",
    "python-dotenv>=1.1.0",
]
